Kyle Clean and Merge Data CSV
Wanted to help out knowing that the weekends weren't the best for you all.
I did not git add/push or merge/send it to my branch (was having a rough time at it). This was not saved on our project_3 vsc repo file. This is a file saved on a separate notebook and I uploaded manually on GitHub, so it shouldn't affect any work that may have happened. If we like how this csv looks and think we can use it to pull what we need, we can copy the code over to .